refactor: move logging utility out of const module
