+++ solver/core/logging.py
@@ -1,0 +1,5 @@
+"""Logging utility for the site."""
+
+import logging
+
+logger = logging.getLogger("uvicorn.error")
